+++ CHANGELOG.md
@@ -6,6 +6,7 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 ## [Unreleased]
 ### Added
 - New column `buffered_at` on the stream_buffer tables to allow for monitoring of stuck stream_buffer events
+- New integration test for event error handling
 ### Changed
 - Jmx MBean `SystemCommanderMBean` now only takes basic Java Objects to keep the JMX handling interoperable
 ### Removed